requirements.txt
absl-py==1.0.0; python_version >= "3.6"
aiohttp==3.8.1; python_version >= "3.6"
aiosignal==1.2.0; python_version >= "3.6"
astunparse==1.6.3
async-timeout==4.0.2; python_version >= "3.6"
attrs==21.4.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
cachetools==5.0.0; python_version >= "3.7" and python_version < "4.0" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
certifi==2021.10.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0"
charset-normalizer==2.0.10; python_full_version >= "3.6.0" and python_version >= "3.6"
clearml==1.1.6
click==8.0.3; python_version >= "3.6"
colorama==0.4.4; python_version >= "3.6" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.6" and python_full_version >= "3.5.0"
cycler==0.11.0; python_version >= "3.7"
datasets==1.18.0
dill==0.3.4; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.1.0"
filelock==3.4.2; python_version >= "3.7" and python_full_version >= "3.6.0"
flatbuffers==2.0
fonttools==4.29.0; python_version >= "3.7"
frozenlist==1.3.0; python_version >= "3.7"
fsspec==2022.1.0; python_version >= "3.6"
furl==2.1.3
future==0.18.2; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
gast==0.4.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
google-auth-oauthlib==0.4.6; python_version >= "3.6"
google-auth==2.4.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
google-pasta==0.2.0
grpcio==1.43.0; python_version >= "3.6"
h5py==3.6.0; python_version >= "3.7"
huggingface-hub==0.4.0; python_full_version >= "3.6.0"
idna==3.3; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
importlib-metadata==4.10.1; python_version < "3.10" and python_version >= "3.7"
importlib-resources==5.4.0; python_version < "3.9" and python_version >= "3.7"
joblib==1.1.0; python_version >= "3.7"
jsonschema==4.4.0; python_version >= "3.7"
keras-preprocessing==1.1.2
keras==2.7.0
kiwisolver==1.3.2; python_version >= "3.7"
libclang==11.1.0
markdown==3.3.6; python_version >= "3.6"
matplotlib==3.5.1; python_version >= "3.7"
multidict==6.0.2; python_version >= "3.7"
multiprocess==0.70.12.2
numpy==1.22.1
oauthlib==3.1.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
opt-einsum==3.3.0; python_version >= "3.5"
orderedmultidict==1.0.1
packaging==21.3; python_version >= "3.7" and python_full_version >= "3.6.0"
pandas==1.4.0; python_version >= "3.8"
pathlib2==2.3.6
pillow==9.0.0; python_version >= "3.7"
protobuf==3.19.3; python_version >= "3.6"
psutil==5.9.0; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
pyarrow==6.0.1; python_version >= "3.6"
pyasn1-modules==0.2.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
pyasn1==0.4.8; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") or python_full_version >= "3.6.0" and python_version >= "3.6" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
pyjwt==2.1.0; python_version >= "3.6"
pyparsing==3.0.7; python_version >= "3.7"
pyrsistent==0.18.1; python_version >= "3.7"
python-dateutil==2.8.2; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
pytz==2021.3; python_version >= "3.8"
pyyaml==6.0; python_version >= "3.6" and python_full_version >= "3.6.0"
requests-oauthlib==1.3.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
requests==2.27.1; python_full_version >= "3.6.0" and python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
rsa==4.8; python_version >= "3.6" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
scikit-learn==1.0.2; python_version >= "3.7"
scipy==1.7.3; python_version >= "3.7" and python_version < "3.11"
seaborn==0.11.2; python_version >= "3.6"
setuptools-scm==6.4.2; python_version >= "3.7"
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
tensorboard-data-server==0.6.1; python_version >= "3.6"
tensorboard-plugin-wit==1.8.1; python_version >= "3.6"
tensorboard==2.8.0; python_version >= "3.6"
tensorflow-addons==0.15.0
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1; python_version >= "3.7" and python_version < "3.11"
termcolor==1.1.0
threadpoolctl==3.0.0; python_version >= "3.7"
tomli==1.2.3; python_version >= "3.7"
tqdm==4.62.3; python_full_version >= "3.6.0"
typeguard==2.13.3; python_full_version >= "3.5.3"
typer==0.4.0; python_version >= "3.6"
typing-extensions==4.0.1; python_version >= "3.6" and python_full_version >= "3.6.0"
urllib3==1.26.8; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4"
werkzeug==2.0.2; python_version >= "3.6"
wrapt==1.13.3; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
xxhash==2.0.2; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
yarl==1.7.2; python_version >= "3.6"
zipp==3.7.0; python_version < "3.9" and python_version >= "3.7"
